# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/v1/preprocess/resize": {"post": {"tags": ["preprocessing"], "summary": "Resize Image", "description": "Resize an image.\n\nArgs:\n    params: The resize parameters.\n    img_in: The image to resize.\n\nReturns:\n    The resized image.\n\nSources:\n    https://docs.opencv.org/4.x/da/d54/group__imgproc__transform.html#ga47a974309e9102f5f08231edc7e7529d", "operationId": "resize_image_api_v1_preprocess_resize_post", "parameters": [{"name": "width", "in": "query", "required": true, "schema": {"type": "integer", "description": "Target width in pixels", "title": "Width"}, "description": "Target width in pixels"}, {"name": "height", "in": "query", "required": true, "schema": {"type": "integer", "description": "Target height in pixels", "title": "Height"}, "description": "Target height in pixels"}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_resize_image_api_v1_preprocess_resize_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/preprocess/crop": {"post": {"tags": ["preprocessing"], "summary": "Crop Image", "description": "Crop an image to a square in the center.\n\nArgs:\n    img_in: The image to crop.\n\nReturns:\n    The cropped image.", "operationId": "crop_image_api_v1_preprocess_crop_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_crop_image_api_v1_preprocess_crop_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/preprocess/grayscale": {"post": {"tags": ["preprocessing"], "summary": "Grayscale Image", "description": "Convert an image to grayscale.\n\nArgs:\n    img_in: The image to convert.\n\nReturns:\n    The grayscaled image.\n\nSources:\n    https://docs.opencv.org/4.x/d8/d01/group__imgproc__color__conversions.html#ga397ae87e1288a81d2363b61574eb8cab", "operationId": "grayscale_image_api_v1_preprocess_grayscale_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_grayscale_image_api_v1_preprocess_grayscale_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/preprocess/noise_reduction": {"post": {"tags": ["preprocessing"], "summary": "Noise Reduct Image", "description": "Apply noise reduction to an image.\n\nArgs:\n    params: The noise reduction parameters.\n    image: The image to process.\n\nReturns:\n    The processed image.\n\nSources:\n    - Gaussian Blur: https://docs.opencv.org/4.x/d4/d86/group__imgproc__filter.html#gaabe8c836e97159a9193fb0b11ac52cf1\n    - Median Blur: https://docs.opencv.org/4.x/d4/d86/group__imgproc__filter.html#ga564869aa33e58769b4469101aac458f9\n    - Bilateral Filter: https://docs.opencv.org/4.x/d4/d86/group__imgproc__filter.html#ga9d7064d478c95d60003cf859260703f", "operationId": "noise_reduct_image_api_v1_preprocess_noise_reduction_post", "parameters": [{"name": "technique", "in": "query", "required": true, "schema": {"$ref": "#/components/schemas/NoiseReductionTechniques", "description": "Noise reduction technique"}, "description": "Noise reduction technique"}, {"name": "kernel_size", "in": "query", "required": false, "schema": {"type": "integer", "description": "Kernel size (odd for median)", "default": 5, "title": "Kernel Size"}, "description": "Kernel size (odd for median)"}, {"name": "sigma_color", "in": "query", "required": false, "schema": {"type": "number", "description": "SigmaColor for bilateral filter", "default": 75.0, "title": "Sigma Color"}, "description": "SigmaColor for bilateral filter"}, {"name": "sigma_space", "in": "query", "required": false, "schema": {"type": "number", "description": "SigmaSpace for bilateral filter", "default": 75.0, "title": "Sigma Space"}, "description": "SigmaSpace for bilateral filter"}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_noise_reduct_image_api_v1_preprocess_noise_reduction_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/preprocess/normalization": {"post": {"tags": ["preprocessing"], "summary": "Normalize Image", "description": "Normalize an image.\n\nArgs:\n    params: The normalization parameters.\n    img_in: The image to normalize.\n\nReturns:\n    The normalized image.\n\nSources:\n    - Rescaling: https://docs.opencv.org/4.x/d2/de8/group__core__array.html#ga87eef7ee3970f86906d1e9d701839f6e\n    - Histogram Equalization: https://docs.opencv.org/4.x/d6/dc7/group__imgproc__hist.html#ga7e54091f0c937d49bf84152a16f76d6e", "operationId": "normalize_image_api_v1_preprocess_normalization_post", "parameters": [{"name": "technique", "in": "query", "required": true, "schema": {"$ref": "#/components/schemas/NormalizationTechniques", "description": "Normalization technique"}, "description": "Normalization technique"}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_normalize_image_api_v1_preprocess_normalization_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/preprocess/binarization": {"post": {"tags": ["preprocessing"], "summary": "Binarize Image", "description": "Binarize an image.\n\nArgs:\n    params: The binarization parameters.\n    image: The image to binarize.\n\nReturns:\n    The binarized image.\n\nSources:\n    https://docs.opencv.org/4.x/d7/d4d/group__imgproc__threshold.html#ga72b913f352e4a1b1b397736707afcde3", "operationId": "binarize_image_api_v1_preprocess_binarization_post", "parameters": [{"name": "technique", "in": "query", "required": true, "schema": {"$ref": "#/components/schemas/BinarizationTechniques", "description": "Binarization technique"}, "description": "Binarization technique"}, {"name": "threshold", "in": "query", "required": false, "schema": {"type": "integer", "description": "Threshold value", "default": 127, "title": "Threshold"}, "description": "Threshold value"}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_binarize_image_api_v1_preprocess_binarization_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/preprocess/contrast": {"post": {"tags": ["preprocessing"], "summary": "Enhance Contrast Image", "description": "Enhance the contrast of an image.\n\nArgs:\n    params: The contrast enhancement parameters.\n    image: The image to process.\n\nReturns:\n    The processed image.\n\nSources:\n    https://docs.opencv.org/4.x/d6/dc7/group__imgproc__hist.html#a5d2d804e716140362f3b9b191447b393", "operationId": "enhance_contrast_image_api_v1_preprocess_contrast_post", "parameters": [{"name": "technique", "in": "query", "required": true, "schema": {"$ref": "#/components/schemas/ContrastTechniques", "description": "Contrast technique"}, "description": "Contrast technique"}, {"name": "clip_limit", "in": "query", "required": false, "schema": {"type": "number", "description": "CLAHE clip limit", "default": 2.0, "title": "Clip Limit"}, "description": "CLAHE clip limit"}, {"name": "tile_grid_size", "in": "query", "required": false, "schema": {"type": "integer", "description": "CLAHE tile grid size", "default": 8, "title": "Tile Grid Size"}, "description": "CLAHE tile grid size"}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_enhance_contrast_image_api_v1_preprocess_contrast_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"BinarizationTechniques": {"type": "string", "enum": ["binary", "binary_inv", "trunc", "tozero", "tozero_inv"], "title": "BinarizationTechniques", "description": "Techniques for image binarization."}, "Body_binarize_image_api_v1_preprocess_binarization_post": {"properties": {"image": {"type": "string", "format": "binary", "title": "Image"}}, "type": "object", "required": ["image"], "title": "Body_binarize_image_api_v1_preprocess_binarization_post"}, "Body_crop_image_api_v1_preprocess_crop_post": {"properties": {"img_in": {"type": "string", "format": "binary", "title": "Img In"}}, "type": "object", "required": ["img_in"], "title": "Body_crop_image_api_v1_preprocess_crop_post"}, "Body_enhance_contrast_image_api_v1_preprocess_contrast_post": {"properties": {"image": {"type": "string", "format": "binary", "title": "Image"}}, "type": "object", "required": ["image"], "title": "Body_enhance_contrast_image_api_v1_preprocess_contrast_post"}, "Body_grayscale_image_api_v1_preprocess_grayscale_post": {"properties": {"img_in": {"type": "string", "format": "binary", "title": "Img In"}}, "type": "object", "required": ["img_in"], "title": "Body_grayscale_image_api_v1_preprocess_grayscale_post"}, "Body_noise_reduct_image_api_v1_preprocess_noise_reduction_post": {"properties": {"image": {"type": "string", "format": "binary", "title": "Image"}}, "type": "object", "required": ["image"], "title": "Body_noise_reduct_image_api_v1_preprocess_noise_reduction_post"}, "Body_normalize_image_api_v1_preprocess_normalization_post": {"properties": {"img_in": {"type": "string", "format": "binary", "title": "Img In"}}, "type": "object", "required": ["img_in"], "title": "Body_normalize_image_api_v1_preprocess_normalization_post"}, "Body_resize_image_api_v1_preprocess_resize_post": {"properties": {"img_in": {"type": "string", "format": "binary", "title": "Img In"}}, "type": "object", "required": ["img_in"], "title": "Body_resize_image_api_v1_preprocess_resize_post"}, "ContrastTechniques": {"type": "string", "enum": ["clahe"], "title": "ContrastTechniques", "description": "Techniques for contrast enhancement."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "NoiseReductionTechniques": {"type": "string", "enum": ["gaussian_blur", "median_blur", "bilateral_filter"], "title": "NoiseReductionTechniques", "description": "Techniques for noise reduction."}, "NormalizationTechniques": {"type": "string", "enum": ["rescaling", "histogram_equalization"], "title": "NormalizationTechniques", "description": "Techniques for image normalization."}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}